Calculator is fully styled and works with decimal numbers

+++ Data/Migrations/20200126183954_CreateCalculationsTable.cs
@@ -13,8 +13,8 @@ namespace MvcCalculator.Data.Migrations
                     Id = table.Column<uint>(nullable: false)
                         .Annotation("Sqlite:Autoincrement", true),
                     UserId = table.Column<string>(nullable: false),
-                    Number1 = table.Column<float>(nullable: false),
-                    Number2 = table.Column<float>(nullable: false),
+                    Number1 = table.Column<decimal>(nullable: false),
+                    Number2 = table.Column<decimal>(nullable: false),
                     Operator = table.Column<string>(nullable: false)
                 },
                 constraints: table =>
